Split rocket and launch actions
