+++ src/CMakeLists.txt
@@ -22,7 +22,8 @@ list(APPEND SolCorpSources
     modules/window/building_detail_window.cpp
     modules/window/notification_window.cpp
     modules/rocket/rocket_module.cpp
-    modules/rocket/actions.cpp
+    modules/rocket/launch_actions.cpp
+    modules/rocket/rocket_actions.cpp
     modules/rocket/launch_window.cpp
     modules/rocket/active_launches_window.cpp
     modules/rocket/contracts_window.cpp
@@ -62,7 +63,8 @@ list(APPEND SolCorpHeaders
     modules/window/building_detail_window.h
     modules/window/notification_window.h
     modules/rocket/rocket_module.h
-    modules/rocket/actions.h
+    modules/rocket/launch_actions.h
+    modules/rocket/rocket_actions.h
     modules/rocket/launch_window.h
     modules/rocket/active_launches_window.h
     modules/rocket/contracts_window.h
