+++ src/modules/rocket/rocket_module.cpp
@@ -1,12 +1,13 @@
 #include "rocket_module.h"
-#include "actions.h"
 #include "active_launches_window.h"
 #include "contracts_window.h"
+#include "launch_actions.h"
 #include "launch_window.h"
 #include "modules/base/action.h"
 #include "modules/base/assert.h"
 #include "modules/base/base.h"
 #include "modules/base/notification.h"
+#include "rocket_actions.h"
 
 #include "modules/engine/engine.h"
 #include "modules/engine/gui.h"
