+++ src/modules/site/rocket_prefab_window.cpp
@@ -3,7 +3,7 @@
 #include "modules/base/assert.h"
 #include "modules/engine/gui.h"
 #include "modules/engine/helpers.h"
-#include "modules/rocket/actions.h"
+#include "modules/rocket/rocket_actions.h"
 #include "modules/rocket/rocket_module.h"
 #include "modules/simulation/simulation.h"
 #include "modules/site/site.h"
