+++ src/modules/window/building_detail_window.cpp
@@ -3,8 +3,8 @@
 #include "modules/base/assert.h"
 #include "modules/base/base.h"
 #include "modules/engine/gui.h"
-#include "modules/rocket/actions.h"
 #include "modules/rocket/launch_window.h"
+#include "modules/rocket/rocket_actions.h"
 #include "modules/rocket/rocket_module.h"
 #include "modules/site/rocket_prefab_window.h"
 #include "modules/site/site.h"
