+++ test/rocket/actions/rocket_complete_build_action.test.cpp
@@ -1,5 +1,5 @@
 #include "modules/base/base.h"
-#include "modules/rocket/actions.h"
+#include "modules/rocket/rocket_actions.h"
 #include "modules/rocket/rocket_module.h"
 #include "modules/simulation/simulation.h"
 #include "modules/window/window_module.h"
