+++ test/rocket/actions/rocket_move_action.test.cpp
@@ -1,4 +1,4 @@
-#include "modules/rocket/actions.h"
+#include "modules/rocket/rocket_actions.h"
 #include "modules/rocket/rocket_module.h"
 #include "modules/simulation/simulation.h"
 #include "modules/window/window_module.h"
