+++ test/rocket/actions/schedule_launch_action.test.cpp
@@ -1,4 +1,4 @@
-#include "modules/rocket/actions.h"
+#include "modules/rocket/launch_actions.h"
 #include "modules/rocket/rocket_module.h"
 #include "modules/simulation/simulation.h"
 #include "modules/site/site.h"
